Move query builder components into own folders
