+++ frontend/src/pages/queryBuilder/textEditor/TextEditor.jsx
