+++ frontend/src/pages/queryBuilder/textEditor/textEditorRow/NodeSelector.jsx
