+++ frontend/src/pages/queryBuilder/textEditor/textEditorRow/PredicateSelector.jsx
