+++ frontend/src/pages/queryBuilder/textEditor/textEditorRow/TextEditorRow.jsx
