+++ frontend/src/pages/queryBuilder/textEditor/textEditorRow/textEditorRow.css
